Allow `POST /populate` requests outside of app
As Rails provides by default CSRF protection for POST requests,
we need to disable it for the `populate` action (for now) as we
intend to (possibly) use it from an API client.

We should consider providing a separate authentication mechanism for
API clients (e.g. API tokens) and re-enabling CSRF protection for
`populate` action.